Fixed-Gruntfile.js must only in root directory
Fixed Known Issue #1- The Gruntfile must be in the root of your project
directory to successfully get the available tasks. Additionally, all
grunt commands will be called in the root directory.
Now, the Gruntfile.coffee (or Gruntfile.js) can be in a sub-directory
of the root directory. And all grunt will run in the Gruntfile
directory.